requirements.txt
openai>=1.0.0
pygame>=2.5.0
python-dotenv>=1.0.0
sounddevice>=0.5.0
scipy>=1.10.0
webrtcvad>=2.0.10
numpy>=1.20.0
langgraph>=0.5.0
langchain-openai>=0.3.0
pvporcupine>=3.0.0
pyaudio>=0.2.14
